# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Meeting {
  id: number;
  title: string;
  /** ISO date (YYYY-MM-DD) */
  meetingDate: string;
  /** Time (HH:MM) */
  meetingTime: string;
  /** @nullable */
  location: string | null;
  participants: string[];
  agenda: string;
  notes: string;
  tags: string[];
  /** One of "scheduled", "completed", "cancelled" */
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMeetingInput {
  /** @minLength 1 */
  title: string;
  meetingDate: string;
  meetingTime: string;
  /** @nullable */
  location?: string | null;
  participants: string[];
  agenda: string;
  notes: string;
  tags?: string[];
  status?: string;
}

export interface UpdateMeetingInput {
  /** @minLength 1 */
  title?: string;
  meetingDate?: string;
  meetingTime?: string;
  /** @nullable */
  location?: string | null;
  participants?: string[];
  agenda?: string;
  notes?: string;
  tags?: string[];
  status?: string;
}

export interface MeetingsSummary {
  total: number;
  scheduled: number;
  completed: number;
  cancelled: number;
  upcomingCount: number;
  uniqueParticipants: number;
}

export interface TagCount {
  tag: string;
  count: number;
}

export type ListMeetingsParams = {
  /**
   * Search by title or participant name
   */
  search?: string;
  /**
   * Filter by tag
   */
  tag?: string;
};