# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Meeting {
  id: number;
  title: string;
  /** ISO date (YYYY-MM-DD) */
  meetingDate: string;
  /** Time (HH:MM) */
  meetingTime: string;
  /** @nullable */
  location: string | null;
  participants: string[];
  agenda: string;
  notes: string;
  tags: string[];
  /** One of "scheduled", "completed", "cancelled" */
  status: string;
  createdAt: Date;
  updatedAt: Date;
}