# lib/api-zod/src/generated/types/meetingsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingsSummary {
  total: number;
  scheduled: number;
  completed: number;
  cancelled: number;
  upcomingCount: number;
  uniqueParticipants: number;
}