# lib/api-zod/src/generated/types/updateMeetingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeetingInput {
  /** @minLength 1 */
  title?: string;
  meetingDate?: string;
  meetingTime?: string;
  /** @nullable */
  location?: string | null;
  participants?: string[];
  agenda?: string;
  notes?: string;
  tags?: string[];
  status?: string;
}